Added Application list and templating support.
Fiddled with the application display. Turns out we can't store the C# objects
like we did in Python. Now storing the ID of the row in the tree rather than
the object.

Fiddled with a couple of templating engines to display the summary of the
application data. Finally settled on StringTemplate. NVelocity has not been
contributed to since 2003 and is grossly difficult to set up for what we
need. StringTemplate is simple.