Default for Login and ControlWrapper for summary.
Just a whole bunch of stuff. First thing was to create the control wrappers
DLL with base classes to support the controls. Then a specific control was
written for the main summary display.

Then, got sick of tabbing through the login controls, so set up the default
button handler.

Also reworked the command-line nant build to be current.