Created renderer and updated domains.
The renderer class is a generic class that knows how to reference the
StringTemplate templates and render them based on the domain name and the
type of template desired (e.g. Summary or Label).

The domain objects were updated to have a property for each of the attributes
and relationships. This will make the domain objects easier to work with in
the templates.