Initial attempt to write data bound controls.
Took the simplest case of an Entry and updated the ApplicationEntryDlg
to use this new control. Got it to work OK, but still need some work. Not
clear that this is the final solution.

Still think that the primary window and dialogs should have a "Context" which
provides a place to hold the data to support the code needed by the attributes.