Corrected the relative template dir handling.
There was some issue with handling a relative template directory. This was
corrected by obtaining the path from the assembly and assuming that the
template directory is relative to that.

There was some funky handling of the assembly path (it was a URL), but other
than that, it was relatively straightforward.